Revert "Rename import"
This reverts commit bf803250d132b967257b9cac0dfb66b5b8f46861.
